Refactor: Add register directory.
