+++ src/app/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import registerReducer from './registerSlice';
+import registerReducer from './register/slice';
 
 export default configureStore({
   reducer: {
